fix: Only install xeus-sql on x86_64

+++ modules/lang.sql.sh
@@ -9,19 +9,20 @@ function as_root() {
     curl -fsSL https://deb.nodesource.com/setup_14.x | bash -
     apt install -y nodejs build-essential
 
-    # Download the SQL Jupyter Kernel (see https://github.com/pojntfx/xeus-sql-binaries#installation)
+    # Download the SQL Jupyter Kernel on x86_64 (see https://github.com/pojntfx/xeus-sql-binaries#installation)
+    if [ "$(uname -m)" = 'x86_64' ]; then
+        # Fetch the xeus-sql binary package
+        curl -L -o /tmp/xeus-sql.tar.gz https://github.com/pojntfx/xeus-sql-binaries/releases/download/latest/xeus-sql.$(uname -m).tar.gz
 
-    # Fetch the xeus-sql binary package
-    curl -L -o /tmp/xeus-sql.tar.gz https://github.com/pojntfx/xeus-sql-binaries/releases/download/latest/xeus-sql.$(uname -m).tar.gz
+        # Extract the package to /usr/local/xeus-sql
+        XEUS_PREFIX=/usr/local/xeus-sql
+        mkdir -p ${XEUS_PREFIX}
+        tar -C ${XEUS_PREFIX} -xzf /tmp/xeus-sql.tar.gz
+        rm /tmp/xeus-sql.tar.gz
 
-    # Extract the package to /usr/local/xeus-sql
-    XEUS_PREFIX=/usr/local/xeus-sql
-    mkdir -p ${XEUS_PREFIX}
-    tar -C ${XEUS_PREFIX} -xzf /tmp/xeus-sql.tar.gz
-    rm /tmp/xeus-sql.tar.gz
-
-    # Install the kernel
-    jupyter kernelspec install ${XEUS_PREFIX}/share/jupyter/kernels/xsql --sys-prefix
+        # Install the kernel
+        jupyter kernelspec install ${XEUS_PREFIX}/share/jupyter/kernels/xsql --sys-prefix
+    fi
 }
 
 # User script
